remove unused serverless functions
