+++ api/routes/health_router.py
@@ -1,11 +1,11 @@
-from fastapi import APIRouter
+# from fastapi import APIRouter
 
-router = APIRouter()
+# router = APIRouter()
 
-@router.get("/health")
-async def health_check():
-    """Simple health check endpoint"""
-    return {
-        'status': 'healthy',
-        'message': 'Code execution service is running'
-    }
+# @router.get("/health")
+# async def health_check():
+#     """Simple health check endpoint"""
+#     return {
+#         'status': 'healthy',
+#         'message': 'Code execution service is running'
+#     }
