+++ types/project/general.ts
@@ -1,0 +1,6 @@
+export interface Progress {
+  completed: boolean
+  totalAttempts: number
+  successfulAttempts: number
+  lastAttempt: number
+}
