+++ types/project/props.ts
@@ -1,0 +1,5 @@
+export interface useProjectProgressOptions {
+  revalidateOnFocus?: boolean
+  revalidateOnReconnect?: boolean
+  dedupingInterval?: number
+}
